# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Admin API
  description: 工作流平台管理后台 REST API
  version: 1.0.0
  contact:
    name: Workflow Platform Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/admin/schemas:
    get:
      operationId: listSchemas
      tags: [Schemas]
      summary: 分页查询 Schema
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: schemaType
          in: query
          schema:
            type: string
          description: 按类型标签过滤（INPUT/OUTPUT/EVENT），仅返回包含该标签的 Schema
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseSchemaDefinition'

    post:
      operationId: createSchema
      tags: [Schemas]
      summary: 注册 Schema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaDefinition'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'
  /api/admin/schemas/{schemaName}:
    get:
      operationId: getSchema
      tags: [Schemas]
      summary: 获取 Schema 详情
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateSchema
      tags: [Schemas]
      summary: 更新 Schema
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaDefinition'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'

    delete:
      operationId: deleteSchema
      tags: [Schemas]
      summary: 删除 Schema
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '204':
          description: 删除成功
  /api/admin/schemas/tables:
    get:
      operationId: listTables
      tags: [Schemas]
      summary: 列出数据库中所有用户表
      parameters:
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 表名列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/admin/schemas/tables/{tableName}/columns:
    get:
      operationId: listTableColumns
      tags: [Schemas]
      summary: 获取指定表的列元数据
      parameters:
        - $ref: '#/components/parameters/TableNamePath'
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 列元数据列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableColumnMetadata'
  /api/admin/schemas/tables/detail:
    get:
      operationId: listTablesWithColumns
      tags: [Schemas]
      summary: 获取所有表及其列信息，用于设计器侧边栏一次性加载
      parameters:
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 表及其列列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableDetail'
  /api/admin/schemas/tables/{tableName}/foreign-keys:
    get:
      operationId: listForeignKeys
      tags: [Schemas]
      summary: 获取指定表的外键信息
      parameters:
        - $ref: '#/components/parameters/TableNamePath'
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 外键信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForeignKeysResponse'
  /api/admin/schemas/tables/{tableName}/json-schema:
    get:
      operationId: generateTableJsonSchema
      tags: [Schemas]
      summary: 根据表结构生成 JSON Schema，不持久化
      parameters:
        - $ref: '#/components/parameters/TableNamePath'
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 生成的 JSON Schema 字符串
          content:
            application/json:
              schema:
                type: string
  /api/admin/schemas/{schemaName}/versions:
    get:
      operationId: listSchemaVersions
      tags: [Schemas]
      summary: 查询 Schema 历史版本
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '200':
          description: 版本列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchemaDefinition'

  # ============ Function API ============
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: 页码（从 0 开始）
      schema:
        type: integer
        default: 0
    PageSize:
      name: pageSize
      in: query
      description: 每页条数
      schema:
        type: integer
        default: 20
...
    SchemaNamePath:
      name: schemaName
      in: path
      required: true
      schema:
        type: string
    TableNamePath:
      name: tableName
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
          description: Unix 毫秒时间戳
        status:
          type: integer
        error:
          type: string
        path:
          type: string

    # ============ 枚举类型 ============
...
    TableColumnMetadata:
      type: object
      properties:
        name:
          type: string
          description: 列名
        dataType:
          type: string
          description: 数据库类型名
        sqlType:
          type: integer
          description: java.sql.Types 常量值
        size:
          type: integer
          description: 列长度/精度
        nullable:
          type: boolean
          description: 是否可空
        defaultValue:
          type: string
          description: 默认值
        comment:
          type: string
          description: 列注释
        ordinalPosition:
          type: integer
          description: 列顺序
        autoIncrement:
          type: boolean
          description: 是否自增
    TableDetail:
      type: object
      properties:
        name:
          type: string
          description: 表名
        columns:
          type: array
          items:
            $ref: '#/components/schemas/TableColumnMetadata'
          description: 列元数据列表
    ForeignKeyInfo:
      type: object
      properties:
        fkName:
          type: string
          nullable: true
          description: 外键约束名
        sourceTable:
          type: string
          description: 来源表名
        sourceColumn:
          type: string
          description: 来源列名
        targetTable:
          type: string
          description: 目标表名
        targetColumn:
          type: string
          description: 目标列名
        keySeq:
          type: integer
          description: 外键列顺序
    ForeignKeysResponse:
      type: object
      properties:
        incoming:
          type: array
          items:
            $ref: '#/components/schemas/ForeignKeyInfo'
          description: 指向当前表的外键
        outgoing:
          type: array
          items:
            $ref: '#/components/schemas/ForeignKeyInfo'
          description: 当前表指向其他表的外键
    SchemaDefinition:
      type: object
      required: [schemaName, schemaType, schemaJson]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        schemaName:
          type: string
          minLength: 1
          maxLength: 256
          description: Schema 唯一名称
        schemaType:
          type: string
          minLength: 1
          maxLength: 64
          description: 类型标签（逗号分隔，如 INPUT,OUTPUT），支持多类型复用
        schemaFormat:
          type: string
          default: json-schema
          description: Schema 格式标识（如 json-schema、protobuf、avro，支持扩展自定义格式）
        schemaJson:
          type: string
          minLength: 1
          maxLength: 65535
          description: Schema 内容（字符串形式，具体语法由 schemaFormat 决定）
        description:
          type: string
          maxLength: 1024
          nullable: true
        frozen:
          type: boolean
          default: false
          description: 是否已冻结（锁定），冻结后普通用户不可修改/删除
        scope:
          type: string
          maxLength: 16
          default: PLATFORM
          description: 作用域：PLATFORM/PRIVATE
        appGroup:
          type: string
          maxLength: 128
          description: 所属应用分组（scope=PRIVATE 时必填）
        createdAt:
          type: integer
          format: int64
          readOnly: true
          description: Unix 毫秒时间戳
        updatedAt:
          type: integer
          format: int64
          readOnly: true
          description: Unix 毫秒时间戳


    # ============ Function 相关 ============
...
    PageResponseSchemaDefinition:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/SchemaDefinition'
        total:
          type: integer
          format: int64
        page:
          type: integer
        pageSize:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT